[FEAT] 사용자 포트폴리오 및 원화 자산 조회

+++ src/main/java/cowing/auth/entity/User.java
@@ -34,7 +34,7 @@ public class User {
     private Authority authority;
 
     @Column(precision = 20, scale = 8, nullable = true)
-    private BigDecimal uHoldings;
+    private Long uHoldings;
 
     @PrePersist
     public void prePersist() {
